requirements.txt
streamlit
pillow
PyMuPDF
pdf2image
pytesseract
rapidfuzz
langchain
langchain-community
langchain-ollama
llama-index
faiss-cpu
unstructured
